apps/agent-python/requirements.txt
livekit-agents
python-dotenv
google-generativeai
